Changes in Client and DataPacket
Client now implements data sending almost normally and receives Acks. If Crc could not be fitted to the packet, then crc is to sent separtely.
TODO
After that Server needs to be configured to be able to understand how much data should be read from the packets. Then onwards from that we will move on to making a normal connection.